pyproject.toml
[project]
name = "useful-scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aioshutil>=1.6",
    "pyyaml>=6.0.3",
]
